Add Grid Component
