+++ src/App.js
@@ -1,13 +1,9 @@
-import { useState, useEffect, useRef } from "react";
+import { useState, useEffect } from "react";
 import Tile from "./components/Tile";
-import { TileMenu } from "./components/TileMenu";
+import Grid from "./components/Grid";
 import "./App.css";
 
 function App() {
-  const gridEl = useRef(null);
-  const [pressed, setPressed] = useState(false);
-  const [position, setPosition] = useState({ x: null, y: null });
-
   /**
    * Generates a UUID to prevent same key error
    * @returns {string}
@@ -38,41 +34,6 @@ function App() {
       row.forEach((tile) => tile !== 0 && renderCreatables(tile))
     );
   }, []);
-
-  /**
-   * When Grid is loaded into the DOM CSS will have centered the Grid.
-   * We take the position of the Grid and put it into the Position state
-   */
-  useEffect(() => {
-    if (!gridEl.current) return;
-    setPosition({ x: gridEl.current.offsetLeft, y: gridEl.current.offsetTop });
-  }, [gridEl]);
-
-  /**
-   * As soon as the Position state changes.
-   * This code will run and update the left and top CSS properties
-   * on the Grid Element.
-   */
-  useEffect(() => {
-    if (!gridEl.current) return;
-    gridEl.current.style.left = `${position.x}px`;
-    gridEl.current.style.top = `${position.y}px`;
-  }, [position]);
-
-  /**
-   * Event Handlers. As soon as the user pressed up or down the Pressed state will toggle.
-   * If the Pressed state is true and the mouse moves; the Position state will be updated.
-   */
-  const handleMouseDown = () => setPressed(true);
-  const handleMouseUp = () => setPressed(false);
-  const handleMouseMove = (event) => {
-    if (pressed) {
-      setPosition({
-        x: position.x + event.movementX,
-        y: position.y + event.movementY,
-      });
-    }
-  };
 
   /**
    * This function takes a tile {object} and checks where on the grid a creatable tile
@@ -168,16 +129,9 @@ function App() {
   }
 
   return (<>
-    <div
-      className="map"
-      onMouseDown={handleMouseDown}
-      onMouseUp={handleMouseUp}
-      onMouseMove={handleMouseMove}
-    >
-      <div className="grid isometric" ref={gridEl}>
+      <Grid isometric>
         {tiles.map((row) => row.map((tile) => tile !== 0 && <Tile key={tile.id} onTileClick={() => handleTileClick(tile)} tile={tile}/>))}
-      </div>
-    </div>
+      </Grid>
   </>);
 }
